* Adding age months * Changing quantity and units type to float

+++ src/main/java/com/hashini/medicare/dao/implementation/MedicineDAOImpl.java
@@ -96,7 +96,7 @@ public class MedicineDAOImpl implements MedicineDAO {
     }
 
     @Override
-    public void updateUnits(long id, int quantityDifference) {
+    public void updateUnits(long id, float quantityDifference) {
         String sql = "UPDATE medicines SET units = units + ? WHERE medicine_id = ?";
         jdbcTemplate.update(sql, quantityDifference, id);
     }
